Add new service for sre api and refactor
