+++ packages/nextjs/app/api/admin/signin/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import { recoverTypedDataAddress } from "viem";
-import { fetchBuilderData } from "~~/services/database/builders";
+import { fetchBuilderData } from "~~/services/api/sre/builders";
 import { EIP_712_DOMAIN, EIP_712_TYPES__ADMIN_SIGN_IN } from "~~/utils/eip712";
 import { validateSafeSignature } from "~~/utils/safe-signature";
