+++ packages/nextjs/app/api/grants/[grantId]/review/route.tsx
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { EIP712TypedData } from "@safe-global/safe-core-sdk-types";
 import { recoverTypedDataAddress } from "viem";
-import { fetchBuilderData } from "~~/services/database/builders";
+import { fetchBuilderData } from "~~/services/api/sre/builders";
 import { reviewGrant } from "~~/services/database/grants";
 import { EIP_712_DOMAIN, EIP_712_TYPES__REVIEW_GRANT, EIP_712_TYPES__REVIEW_GRANT_WITH_NOTE } from "~~/utils/eip712";
 import { PROPOSAL_STATUS, ProposalStatusType } from "~~/utils/grants";
